holobase_queue:tr_queue now excludes tagged ops
When pushing a new op, you tag it with the version of the document it is
applied to and a hash identifying the user who applied the op.
When creating a queue of transformation holobase_queue:tr_queue can now exclude
the ops from the queue that belongs to the same hash (i.e. user).